msgmerge: don't assign empty string to last comment when merging
When a po file including no last comment is parsed, the last comment
doesn't appear in the result by parsing. So the value of the last
comment specifies nil.
On the other hand, when both two files including no last comment are
merged, the result file by merging them has an empty string("") as
the last comment. But this result shouldn't have the last comment
because it doesn't appear in this result.