GetText::MO: support charsets in lowercase
On my system for some reason my locale is set to pt_BR.utf8, and that
makes Encoding.find() be called with "utf8", which is not how Ruby
represents "UTF-8". This commit will try harder to find a matching
encoding amongts the ones supported natively by Ruby comparing the input
charset name with modified versions of the Ruby encodings.